GS-13598 modified demote API in AdminAPI to return a Future + changed timeout to be long instead of int

+++ xap-core/xap-datagrid/src/main/java/com/gigaspaces/internal/client/spaceproxy/IDirectSpaceProxy.java
@@ -131,6 +131,6 @@ public interface IDirectSpaceProxy extends ISpaceProxy {
 
     ISpaceQuery prepareTemplate(Object template);
 
-    void demoteToBackup(int timeout, TimeUnit unit) throws DemoteFailedException;
+    void demoteToBackup(long timeout, TimeUnit unit) throws DemoteFailedException;
 
 }
